Ignore product pages without READMEs

+++ _scripts/gen-product-pages.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 dashes="---"
 for productDir in $HOME/Code/teragon/* ; do
-  productName=$(basename $productDir)
-  printf "%s\nlayout: product\ntitle: %s\nalias: /p/%s.html\n%s\n\n" "$dashes" "$productName" "$productName" "$dashes" > ${productName}.md
-  cat $productDir/README.md >> ${productName}.md
+  if [ -e "$productDir/README.md" ] ; then
+    productName=$(basename $productDir)
+    printf "%s\nlayout: product\ntitle: %s\nalias: /p/%s.html\n%s\n\n" "$dashes" "$productName" "$productName" "$dashes" > ${productName}.md
+    cat $productDir/README.md >> ${productName}.md
+  fi
 done
